add in cap command to calculate current CAP and integrate into our app
